metarbot: Switch to avwx.rest API source
This API covers more METARs, while the old one provided by api.met.no
unfortunately has significant gaps. Since the new API returns JSON, add
a dependency on the serde crate to deserialize that.

Use irc::client::data::config::Config::get_option over option::get,
because the latter is longer and returns a String, while the former
returns a &str, which is sometimes easier to handle for us.

Modify the bot command API to accept borrows rather than copies of many
of the objects that are re-used across invocations, such as the leader,
the owner and the options parameter.

Explicitly handle 204 errors as "no reporting weather".